requirements.txt
PyYAML
requests
click
